add package diff
If the user is looking at an individual component
Add the package diff between current-tripleo and
component-ci-testing.  This shows the user
what packages have changed and are being tested
by the component line.

If the user is looking at an integration pipeline
the diff between tripleo-ci-testing and current-tripleo
will be displayed

e.g.
https://paste.opendev.org/show/807489/
https://paste.opendev.org/show/807488/

Change-Id: I1a0bb85f115ee9acf44028759895a29262bf1edf