Add CentOS composes promoter role
This patch adds a new role to assist on centos compose-id promotion,
which fetches data from a remote url and create compose-id artifacts
in the provided remote server. It includes a module that expects both
candidate and target labels as parameters of a promotion.
At this moment, the supported promotion is 'latest-compose' to
'tripleo-ci-testing'.

Change-Id: Iac19138dc8354d9c98fb26b7d0298db0a83ce165
Signed-off-by: Douglas Viroel <viroel@gmail.com>