Added promoter specific environment variable
We use the same promoter code for downstream and upstream.
The config is kept at two different location.

This code adds promoter_type as upstream and downstream to
manage both the configs during deployment.

It moves all the env var to a file which can be sourced when
we run dlrn-promoter.sh script to load appropriate config.

It will help to do the deployment and configuration easily.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: If38b8d2486096e95a30dde7d5a6f59109b18dfff