Fix: Add distro info to job results
This patch fix job results to include distro information. With this
info we can properly filter job results in the dashboard.
It also fixes the information being displayed on job results in all
affected dashboards (componente jobs running in multiple distros).

Note that after updating the environment with this change, the old
job information won't be displayed in the dashboard, since the database
needs to be populated with distro information now.

Change-Id: Id0821cc9e0c25719c384d25887bc54bd04545b77
Signed-off-by: Douglas Viroel <viroel@gmail.com>