Updating queries on the cockpit
This patch update cockpit queries to add the RETRY, RETRY_LIMIT,
ERROR and LOST, since those are new zuul error status.
Also add a new option in the pie charts for periodic and check
jobs with the DISK_FULL.
It also update the queries to use the regex for result, so now
it is possible to write the query with where result !~ /SKIPPED|
ERROR|RETRY/ instead of keep repeating and result != SKIPPED and
result != RETRY and so on.

Change-Id: I7b5da7433c7986f9fe190413490644b74ae5ebfa