Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

why do we want to remove it from the periodic lines

no we don't have it in gates but would like to keep it in periodic unless there is some reason why we cannot.

if it is broken do we have a bug for it (rhetorical, looks like it fails on the pcs issue) - so if we need to remove from criteria for a bit, sure but not sure we want to remove completely?

Patch-set: 1
Reviewer: Gerrit User 192 <192@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=-1