Fix test_gerrit_changes unit test
This patch fix "test_gerrit_changes" unit test that is failing due to
an extra key which is expected from gerrit api. That key was removed
expedct data and the test was updated to, instead of comparing that all
keys are equal, to compare that the expected keys are a subset of the
actual gerrit keys.
This avoid future issues with updates in gerrit api response, and will
continue to validate that a minimal set of keys is in gerrit api response.
Note that the keys that are being removed were added here[1] just to
fix the test that was failing in past.

[1] https://review.rdoproject.org/r/c/rdo-infra/ci-config/+/34984

Change-Id: I812eea10c1e2d7f3afd3cac6bb7ed81a12d7dcde
Signed-off-by: Douglas Viroel <viroel@gmail.com>