Refactored a way of handling releases versions
Due to constant changes with supported versions of different
OpenStack releases and systems, it is possible that we're
trying to fetch non-existing links.
Recent removal of CentOS 8 master jobs caused issues with
successfully fetching content, causing breakage of the script.

This patch introduces matrix of supported releases and systems.
It should allow for easier management of dependencies.

The script tries to mitigate possible user's mistakes by polling
right distro and release versions based on the matrix.

Change-Id: I95cbcd770bd396cc18e38c5680bbab92bd10c32c