Add cron job to copy containers from quay to rdo
This is a safeguard for when promotion takes longer than five days to
promote, since rdo registry prune old containers every five days.
This patch adds a cron job to copy the containers from quay.io to rdo
registry for all supported releases.
This also add a cleanup on the copy-quay script to delete the build
directory in /tmp in order to avoid the toolbox disk to be full

Change-Id: I75a1242136d1bf8ce8dfc0a6f65cc9f11a660527