Reduce centos8 wallaby promotion criteria
With [1] we removed most of centos-8 wallaby gate across
TripleO. This does the same for the promotion criteria - we keep
the vanilla standalone and featureset 1 and 20 ovb with the
required image and container build jobs.

For components this also adds some missing standalone jobs and
also some of the featureset1 jobs but with comment as they
are unstable. Related patch at [2] depends-on this and removes
these jobs from the periodic lines.

[1] https://review.opendev.org/q/topic:c8_wallaby_teardown
[2] I14f13a0d1b24d6466d8adef68b3ea5b239742517

Change-Id: I9aa1730a91755c985f629b4a4de0e78ab7108a58