Reduce centos8 wallaby promotion criteria
With [1] we removed most of centos-8 wallaby gate across
TripleO. This does the same for the promotion criteria - we keep
the vanilla standalone and featureset 1 and 20 ovb with the
required image and container build jobs.

For components this also adds some missing standalone jobs and
also some of the featureset1 jobs but with comment as they
are unstable. Related patch at [2] depends-on this and removes
these jobs from the periodic lines.

The criteria for current-tripleo-rdo is adjusted according to [3]

[1] https://review.opendev.org/q/topic:c8_wallaby_teardown
[2] https://review.rdoproject.org/r/c/rdo-jobs/+/41831
[3] https://review.rdoproject.org/r/c/rdo-infra/ci-config/+/41856/2/jenkins-v2/jobs/promote-wallaby-centos8-current-tripleo.yml

Change-Id: I9aa1730a91755c985f629b4a4de0e78ab7108a58