Refactoring compose promoter role
This patch refactors compose_promoter role to work as a generic
file promoter that can be reused for other purposes.
The role and module has been renamed to "artifact_promoter".
This python module now supports two types of promotion:
1. 'file' type promotion receives a file name and content and
a target label. Promoter code creates artifacts in remote server
and links to target label.
2. 'centos-compose' for a given compose_url, the promoter will
fetch latest-compose-id and create artifacts on remote server,
and link to a target-label.
Molecule job was updated to test both modes.

Change-Id: Iba486585b39c06fe9dbda0625125f6e4449b5aee
Signed-off-by: Douglas Viroel <viroel@gmail.com>