Add all templates to Docker image
Image which is built, is used only to cooperate with influxDB.
However, it might be also used as a proxy to work with RR script.

RDO servers are located in North America, hence execution times
from other parts of the world take significant time penalty.
To allow every team member to see similar speed ups like in
North America, we need to include all templates in the image.
This is going to allow users to execute "ruck_rover" script like
on local instance, bypassing geo locations.

Change-Id: I307f265423a0337cc51047083b3346fdc9f8a772