Run ansible-pull once a day instead of every 5 minutes
On each run we update several git repositories and update our system
packages. Doing it every minute is overkill since package updates
are provided in much longer intervals only. The only reason why we
trigger it every 5 minutes is that we do not want to wait so long
until our changes are applied to the machines. But we actually (have
to) test our changes in "production" and thus apply our changes
manually on the affected systems anyway. One reason why we apply
changes manually is because changes to containers will only be
picked up when docker-compose.yml file changes but not when e.g.
Dockerfile files change.

If you really need an update to be applied immediately you can
simply ssh to a server and run ansible_pull.sh manually which
will do the same as the cron job. It is safe to do so because
the script aquires a lock before running ansible-pull, so it
will not interfere with running cron jobs.

Change-Id: I033f12934d1df0d57d83e24ccf422cfeed35a826