Fixing wrong hackmd api path
background story:  Looks like hackmd has a bug (or feature?) There are 2 types of notes - personal notes for which the api patch is ...notes/<note_id> and team notes for which patch is .../teams/<team_id>/notes/<note_id> The bot used to use the syntax /notes/ to update the team notes. It worked well for so many months. Since a few weeks this api call returns 403 but still updates the actual note somehow! So the reviews get added to
the list but the bot thinks 403 means it couldnt add reviews to the list

Change-Id: I932adc4a82d415db8b9773c86f9a0f36d0bc8b11