Promoter: Replace criteria with alternative job
Some jobs are run in two different
environments to workaround environment failures.

If there is a failing job, this patch adds logic to
check if there are alternative jobs in the criteria
file, and if any alternative job pass.
If so, it removes the failing job from criteria
and adds the first passing job from an alternative
environment.

Change-Id: Iadf76a71efd897fa2088e11bbf61f805f1259ec0