Prepare support for multiple OpenStack projects
Our Ansible playbooks, inventory and roles for our infrastructure on
Vexxhost are tightly coupled which prevents us from set up a staging
environment. This patch is the first in a series to decouple Ansible
playbooks and roles from the inventory which will allow us to deploy
our infrastructure on several different OpenStack projects.

In this patch a new variable cloud is introduced which will be used
in later patches to choose a suitable OpenStack project config.

Change-Id: I3bce238949f735097cab810829776cfbd67e7bd8