Mark Grafana container as default server in nginx configuration
When connecting to our cockpit servers by ip instead of hostname,
then in the past the user would see a nginx help page instead of
being shown our Grafana dashboards. This has been changed now to
always show the Grafana dashboards to the user when a cockpit is
connected to by ip.

Co-Authored-By: Ananya Banerjee <anbanerj@redhat.com>

Change-Id: I5395f307f6c039e2a8647904979cdb20c93225fa