Updating the puppet dashboard to match the new grafana format
The puppet dashboard was missing data. On inspection it looks like the dashboard is using "Grafana" as datasourse instead of the telegraf query. This might be due to some version mismatch between the json format and grafana.
So in this patch I changed to the correct query on the production dashboard and have exported the json that it produced.

Change-Id: I31bc52a5f1f20ad607296d43e5fff6cbf84d6c69