Adds bobcat panel to puppet dashboard
The initial patch [1] added a bobcat panel to the puppet dashboard. However the scripts which deploy grafana setup on local are not deploying the exact same version of grafana/telegraf as in the production server. As a result the json produced my updating panels on local copy of the dashboard breaks in production.

Here we tried to update the panel directly on the production cockpit (upstream) and exported the json.

[1] https://review.rdoproject.org/r/c/rdo-infra/ci-config/+/48196

Change-Id: I3414804c004a061e3a83a83cf6d3d4f14090f743