Pin docker-ce version - workaround
This patch pinned the docker-ce version to 3:23.0.5-1.el8 (old)
as the new version of docker-ce:3:24.0.0-1.el8 causing molecule
job failure [1]

[1]: https://logserver.rdoproject.org/79/48679/3/check/tripleo-ci-promotion-staging-single-pipeline-centos-8/4cbe285/job-output.txt

Note: This is workaround to fix the molecule jobs failure to unblock the
patches. In the follow up patch will need to check and fix the root
cause of failure with the newer version of docker-ce.

Change-Id: I42532b4b49a90185e311dafd60bfe96696363102