Adapt promoter service to implement KerberosAuth.
Adapts the promoter service to use or not KerberosAuth as implemented in dlrnapi_client.
In order to use kerberosAuth, it's necessary to:
* Have a valid keytab. And execute a kinit command before executing the promoter service.
* Set dlrnauth_auth_method to "kerberosAuth".
* Set dlrnauth_server_principal to the service's kerberos principal.

Also we've added dlrnauth_force_auth in order to add the auth header while using GET http's action.

By default, dlrnauth_auth_method is "basicAuth", dlrnauth_server_principal is None and dlrnauth_force_auth is False.

Change-Id: I93c776012efa823cbcd23b0a758fe2566fece08a